==> notebooks/1.1_Scrapper_Tweets.ipynb <==
# # ruta="/content/drive/My Drive/Noticias/Alertas Negativas/"
# ruta="/content/drive/My Drive/Alertas Negativas/"

!pip install -q twitterscraper

import twitterscraper
from twitterscraper.query import query_tweets #obtener todos los tweets con una palabra clave
import datetime as dt ## para el manejo de las fechas
import time ## para calcular el tiempo de procesamiento
import pandas as pd ### para el manejo de datframes
import numpy as np

import os
os.chdir("/content/drive/My Drive/Colab Notebooks/Alertas_Negativas_Twitter/notebooks")

### Scrapear información de los tuits

today = dt.date.today()
tomorrow = dt.date.today() + dt.timedelta(days=1)

def get_tweet(query,limit=None,begindate=today,enddate=tomorrow,poolsize=20,lang=" "):
  list_tweet=[]
  if enddate != tomorrow:
    enddate=enddate + dt.timedelta(days=1)
  list_text=query_tweets(query, limit=limit,begindate=begindate ,enddate=enddate, poolsize=poolsize, lang=lang) ## es 
  hour_per = dt.timezone(dt.timedelta(hours=-5))#Configurar la zona horaria 
  for tweet in list_text:
        dict_ = {# user name & id
                "screen_name":tweet.screen_name,
                "username":tweet.username,
                "user_id":tweet.user_id,
                # tweet basic data
                "tweet_id":tweet.tweet_id,
                "tweet_url":tweet.tweet_url,
                "timestamp":tweet.timestamp.astimezone(hour_per),
                "timestamp_epochs":tweet.timestamp_epochs,
                # tweet text
                "text":tweet.text,
                "text_html":tweet.text_html,
                "links":tweet.links,
                "hashtags":tweet.hashtags,
                # tweet media
                "has_media":tweet.has_media,
                "img_urls":tweet.img_urls,
                "video_url":tweet.video_url,
                # tweet actions numbers
                "likes":tweet.likes,
                "retweets":tweet.retweets,
                "replies":tweet.replies,
                "is_replied":tweet.is_replied,
                # detail of reply to others
                "is_reply_to":tweet.is_reply_to,
                "parent_tweet_id":tweet.parent_tweet_id,
                "reply_to_users":tweet.reply_to_users
                }
        list_tweet.append(dict_)
  df=pd.DataFrame(list_tweet).sort_values(by="timestamp",ascending=False)
  df.drop_duplicates(subset=["user_id","tweet_id"],keep="first",inplace=True)
  df.index=[i for i in range(0,len(df))]
  return df

tomorrow = dt.date.today() + dt.timedelta(days=1)
df_tweets_total=get_tweet(query="@bcrpoficial",begindate=dt.date(2020,1, 18),enddate=tomorrow,poolsize=150)
display(df_tweets_total)

df_tweets_total.to_pickle("../data/raw/df_tweets_total.pkl")

# df_tweets_total=pd.read_pickle("../data/raw/df_tweets_total.pkl")

##### Obtener todas las publicaciones realizadas por el banco


aux=df_tweets_total.groupby(["parent_tweet_id"]).tweet_id.count().to_frame().rename(columns={"tweet_id":"comment_total"})
df_tweets_banco=df_tweets_total[(df_tweets_total["screen_name"]=="bcrpoficial") & (df_tweets_total["parent_tweet_id"]=="")]
df_tweets_banco.index=[ i for i in range(0,len(df_tweets_banco))]
df_tweets_banco=pd.merge(df_tweets_banco, aux ,how="left",left_on="tweet_id",right_on="parent_tweet_id")
df_tweets_banco["comment_total"]=df_tweets_banco["comment_total"].fillna(value=0).astype("int32")
display(df_tweets_banco)

df_tweets_banco.to_pickle("../data/raw/df_tweets_banco.pkl")

##### obtener todas las respuestas del Banco

df_tweets_banco_repply=df_tweets_total[(df_tweets_total["screen_name"]=="bcrpoficial") & (df_tweets_total["parent_tweet_id"]!="")]
print(df_tweets_banco_repply.shape)
df_tweets_banco_repply.head()

df_tweets_banco_repply.to_pickle("../data/raw/df_tweets_banco_repply.pkl")

##### Obtener publicaciones que mencionan al banco 

### Obtener todas las menciones al Banco Central
df_tweets_mention=df_tweets_total[(df_tweets_total["screen_name"]!="bcrpoficial") & (df_tweets_total["parent_tweet_id"]=="")]
df_tweets_mention.index=[ i for i in range(0,len(df_tweets_mention))]
df_tweets_mention=pd.merge(df_tweets_mention, aux ,how="left",left_on="tweet_id",right_on="parent_tweet_id")
df_tweets_mention["comment_total"]=df_tweets_mention["comment_total"].fillna(value=0).astype("int32")
display(df_tweets_mention)

df_tweets_mention.to_pickle("../data/raw/df_tweets_mention.pkl")

##### Obtener tuits que responden a publicaciones , ya sea del Banco o de terceros

df_tweets_repply=df_tweets_total[(df_tweets_total["screen_name"]!="bcrpoficial") & (df_tweets_total["parent_tweet_id"]!="")]
df_tweets_repply.index=[ i for i in range(0,len(df_tweets_repply))]
display(df_tweets_repply)

df_tweets_repply.to_pickle("../data/raw/df_tweets_repply.pkl")
