==> notebooks/1.2_Scrapper_Followers.ipynb <==
!pip install -q twint
### este script se correria de vez en cuando y formaria como un input para para la demo

import twint  

import os
os.chdir("/content/drive/My Drive/Colab Notebooks/Alertas_Negativas_Twitter/notebooks")

c = twint.Config()
c.Username = "bcrpoficial" ### establecer el screen_name del usuario
c.Store_csv = True ### guardar la info en csv
c.Output = "../data/raw/df_usernames_followers.csv"### nombre de la carpeta a almacenar los datos
c.Followers = True ### para extraer la info de los seguidores
twint.run.Followers(c) ### para obtener la información de los seguidores

c = twint.Config()
c.Username = "bcrpoficial" ### establecer el screen_name del usuario
c.Store_csv = True ### guardar la info en csv
c.Output = "../data/raw/df_usernames_following.csv"### nombre de la carpeta a almacenar los datos
c.Following = True ### para extraer la info a quienes sigue la cuenta
twint.run.Following(c) ### para obtener la informacion de los seguidos
