==> notebooks/2_Scrapper_UserInfo.ipynb <==
### Importar librerias

!pip install -q twitterscraper
!pip install -q twint

from twitterscraper.query import query_user_info #informacion de usuarios
import datetime as dt ## para el manejo de las fechas
from multiprocessing import Pool #procesamiento paralelo
import pandas as pd ### para el manejo de datframes
import numpy as np
import twint

import os
os.chdir("/content/drive/My Drive/Colab Notebooks/Alertas_Negativas_Twitter/notebooks")

### Scrapear información de los usuarios - TwitterScrapper

def get_user_info(twitter_user):
    try:
      user_info = query_user_info(user= twitter_user)
      twitter_user_data = {}
      twitter_user_data["screen_name"] = user_info.user
      twitter_user_data["username"] = user_info.full_name
      twitter_user_data["location"] = user_info.location
      twitter_user_data["blog"] = user_info.blog
      twitter_user_data["date_joined"] = user_info.date_joined
      twitter_user_data["user_id"] = user_info.id
      twitter_user_data["num_tweets"] = user_info.tweets
      twitter_user_data["following"] = user_info.following
      twitter_user_data["followers"] = user_info.followers
      twitter_user_data["likes"] = user_info.likes
      twitter_user_data["lists"] = user_info.lists
    except:
      twitter_user_data="None"
    return twitter_user_data

users=pd.read_pickle("../data/raw/df_tweets_total.pkl")["screen_name"].unique().tolist()
users[0:10]

twitter_user_info=[]
pool = Pool(40)    
for user in pool.map(get_user_info,users):
  if user!="None":
    twitter_user_info.append(user)
df_user_info = pd.DataFrame(twitter_user_info)
display(df_user_info)

### configurando data
meses={"Jan":1,"Feb":2,"Mar":3,"Apr":4,"May":5,"Jun":6,"Jul":7,"Aug":8,"Sep":9,"Oct":10,"Nov":11,"Dec":12}
def replace(x):
  val=x.split("-")[1].split()[1]
  x=x.split("-")[1].replace(val,str(meses[val])).strip().replace(" ","/")
  return x
aux2=df_user_info["date_joined"].apply(lambda x : replace(x)).apply(lambda x : dt.datetime.strptime(x, '%d/%m/%Y'))
df_user_info["date_joined"]=aux2
df_user_info.sort_values(by="date_joined",ascending=False,inplace=True)
df_user_info=df_user_info.set_index("screen_name").reset_index().copy()

df_user_info.drop(labels="lists",axis=1,inplace=True)
df_user_info

df_user_info.to_pickle("../data/raw/df_user_info.pkl")

### Scrapear Informacion de los usuario -- TWINT

users
c = twint.Config()
c.Store_csv = True
c.Output = "../data/raw/df_user_info_twint.csv"
c.User_full=True
for i in users:
  c.Username = i
  twint.run.Lookup(c)

dict_={"username":"screen_name","name":"username","location":"location","url":"blog",
        "join_date":"date_joined","id":"user_id","tweets":"num_tweets","following":"following",
        "followers":"followers","likes":"likes"}
df_user_info_twint=pd.read_csv("../data/raw/df_user_info_twint.csv",sep=",")
df_user_info_twint=df_user_info_twint[list(dict_.keys())]
df_user_info_twint.rename(columns=dict_,inplace=True)
print(df_user_info_twint.shape)
df_user_info_twint.head()

## configurando data a datetime
meses={"Jan":1,"Feb":2,"Mar":3,"Apr":4,"May":5,"Jun":6,"Jul":7,"Aug":8,"Sep":9,"Oct":10,"Nov":11,"Dec":12}
def replace(x):
  val=x.split(" ")[1]
  x=x.replace(val,str(meses[val])).strip().replace(" ","/")
  return x
aux2=df_user_info_twint["date_joined"].apply(lambda x : replace(x)).apply(lambda x : dt.datetime.strptime(x, '%d/%m/%Y'))
df_user_info_twint["date_joined"]=aux2
print(df_user_info_twint.shape)
df_user_info_twint.head()

### UserInfo Final

df_user_info_final=pd.concat([df_user_info,df_user_info_twint],axis=0)
df_user_info_final.drop_duplicates(subset="screen_name",inplace=True)
df_user_info_final

### cargar los seguidores de la cuenta
df_usernames_followers=pd.read_csv("../data/raw/df_usernames_followers.csv",sep=",").rename(columns={"username":"screen_name"})
df_usernames_followers["is_followers"]=[1 for _ in range(0,len(df_usernames_followers))]
print(df_usernames_followers.shape)
df_usernames_followers.head()

###  cargar a quienes sigue la cuenta
df_usernames_following=pd.read_csv("../data/raw/df_usernames_following.csv",sep=",").rename(columns={"username":"screen_name"})
df_usernames_following["is_following"]=[1 for _ in range(0,len(df_usernames_following))]
print(df_usernames_following.shape)
df_usernames_following.head()

### unir las tres bases, df_user_info_final + df_usernames_followers + df_usernames_following
df_user_info_final=pd.merge(df_user_info_final,df_usernames_followers,how="left",on="screen_name")
df_user_info_final=pd.merge(df_user_info_final,df_usernames_following,how="left",on="screen_name")
print(df_user_info_final.shape)
df_user_info_final.head()

### rellenar los campos vacios  nan con 0
df_user_info_final["is_followers"].fillna(0,inplace=True)
df_user_info_final["is_followers"]=df_user_info_final["is_followers"].astype("int32")
df_user_info_final["is_following"].fillna(0,inplace=True)
df_user_info_final["is_following"]=df_user_info_final["is_following"].astype("int32")
# df_user_info_follow["is_followers"]=df_user_info_follow["is_followers"].map({True:1,False:0})
print(df_user_info_final.shape)
df_user_info_final.head()

df_user_info_final.to_pickle("../data/raw/df_user_info_final.pkl")
