==> notebooks/3_Identify_Influencers.ipynb <==
import pandas as pd
import math

import os
os.chdir("/content/drive/My Drive/Colab Notebooks/Alertas_Negativas_Twitter/notebooks")

df_user_info_final=pd.read_pickle("../data/raw/df_user_info_final.pkl")
df_tweets_total=pd.read_pickle("../data/raw/df_tweets_total.pkl")
df_tweets_total=df_tweets_total[df_tweets_total["screen_name"]!="bcrpoficial"].copy()
df_tweets_total.shape

df_tweets_counts=df_tweets_total.groupby(by="screen_name").text.count().to_frame().sort_values("text",ascending=False)
df_tweets_counts.rename(columns={"text":"tweets_counts"},inplace=True)

df_retweets_counts=df_tweets_total.groupby(by="screen_name").retweets.apply(lambda x : sum(x)).to_frame().sort_values("retweets",ascending=False).reset_index()
df_retweets_counts.rename(columns={"retweets":"retweets_counts"},inplace=True)

df_likes_counts=df_tweets_total.groupby(by="screen_name").likes.apply(lambda x : sum(x)).to_frame().sort_values("likes",ascending=False).reset_index()
df_likes_counts.rename(columns={"likes":"likes_counts"},inplace=True)

df_replies_counts=df_tweets_total.groupby(by="screen_name").replies.apply(lambda x : sum(x)).to_frame().sort_values("replies",ascending=False).reset_index()
df_replies_counts.rename(columns={"replies":"replies_counts"},inplace=True)

aux=pd.merge(pd.merge(pd.merge(df_tweets_counts,df_retweets_counts,how="inner",on="screen_name")\
                      ,df_likes_counts,how="inner",on="screen_name")\
                      ,df_replies_counts,how="inner", on="screen_name")
print(aux.shape)
aux.head()

df_user_info_influencers=pd.merge(df_user_info_final,aux,how="inner",on="screen_name")
print(df_user_info_influencers.shape)
df_user_info_influencers.head()

features=["followers","tweets_counts","likes_counts","retweets_counts","replies_counts"]
gains=[]
for ft in features:
  sum_total=df_user_info_influencers[ft].sum()
  gain=-sum([(i/sum_total)*math.log2(i/sum_total) for i in df_user_info_influencers[ft] if i!=0])
  gains.append(gain)
df_user_info_influencers["metrics"]=df_user_info_influencers["followers"]*gains[0]+df_user_info_influencers["tweets_counts"]*gains[1]+df_user_info_influencers["likes_counts"]*gains[2]+df_user_info_influencers["retweets_counts"]*gains[3]+df_user_info_influencers["replies_counts"]*gains[4]

gains

df_user_info_influencers.sort_values(by="metrics",ascending=False,inplace=True)
df_user_info_influencers.index=[i for i in range(0,len(df_user_info_influencers))]
df_user_info_influencers.head()

df_user_info_influencers.to_pickle("../data/intermediate/df_user_info_influencers.pkl")
